TradesProcessor/requirements.txt
pyspark==3.3.2
kafka-python
avro-python3